+++ app/src/main/java/com/example/nutrition_assessment_system_android_app/data/user/repository/UserRepositoryImpl.kt
@@ -136,7 +136,7 @@ class UserRepositoryImpl @Inject constructor(
     override suspend fun updateUserProfile(
         gender: String,
         age: Int,
-        height: Float,
+        height: Int,
         weight: Float,
         activityLevel: String,
         goal: String
